refactor: rename dynamic enum subpackage to 'definitions'
Renamed 'type' subpackage to 'definitions' in dataaccess packages for
PriceType and OrganizationType to better reflect the contents (dynamic
enum value definitions). Updated all imports and package declarations.
