+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/definitions/OrganizationTypeDefinition.java
@@ -1,4 +1,4 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.type;
+package io.commercestacksolutions.priceproviderservice.dataaccess.organization.definitions;
 
 public interface OrganizationTypeDefinition {
     OrganizationType getOrganizationType();
