+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/pricerow/definitions/PriceTypeDefinition.java
@@ -1,4 +1,4 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type;
+package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions;
 
 public interface PriceTypeDefinition {
     PriceType getPriceType();
