+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/pricerow/definitions/RentalBasePriceType.java
@@ -1,4 +1,4 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type;
+package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions;
 
 import org.springframework.stereotype.Component;
