+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/service/pricerow/validation/PriceRowValidPriceTypeRule.java
@@ -3,7 +3,7 @@ package io.commercestacksolutions.priceproviderservice.service.pricerow.validati
 import io.commercestacksolutions.commons.service.entity.validation.ValidationRule;
 import io.commercestacksolutions.commons.web.rest.Message;
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.entity.PriceRowEntity;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceTypeRegistry;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceTypeRegistry;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
