+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/service/publicprice/PublicPriceServiceImpl.java
@@ -1,7 +1,7 @@
 package io.commercestacksolutions.priceproviderservice.service.publicprice;
 
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.entity.PriceRowEntity;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
 import io.commercestacksolutions.priceproviderservice.service.group.GroupHierarchyService;
 import io.commercestacksolutions.priceproviderservice.service.group.model.GroupWithDistance;
 import io.commercestacksolutions.priceproviderservice.service.publicprice.model.PriceMatchingCriteria;
