+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/service/publicprice/strategy/PriceCandidatesQueryParams.java
@@ -1,6 +1,6 @@
 package io.commercestacksolutions.priceproviderservice.service.publicprice.strategy;
 
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
 import io.commercestacksolutions.priceproviderservice.service.group.model.GroupWithDistance;
 
 import java.math.BigDecimal;
