+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/web/controller/publicapi/PublicPriceController.java
@@ -3,8 +3,8 @@ package io.commercestacksolutions.priceproviderservice.web.controller.publicapi;
 import io.commercestacksolutions.commons.exception.InvalidParameterException;
 import io.commercestacksolutions.commons.exception.NotFoundException;
 import io.commercestacksolutions.commons.mapper.exception.DataMappingException;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceType;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceTypeRegistry;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceTypeRegistry;
 import io.commercestacksolutions.priceproviderservice.config.security.JwtClaimsExtractor;
 import io.commercestacksolutions.priceproviderservice.facade.publicprice.PublicPriceFacade;
 import io.commercestacksolutions.priceproviderservice.facade.publicprice.restentity.PublicPriceListRestEntity;
