+++ service/src/test/java/io/commercestacksolutions/priceproviderservice/web/controller/OrganizationControllerTest.java
@@ -4,7 +4,7 @@ import io.commercestacksolutions.commons.exception.DataIntegrityException;
 import io.commercestacksolutions.commons.exception.EntityAlreadyExistsException;
 import io.commercestacksolutions.commons.service.entity.validation.exception.EntityValidationException;
 import io.commercestacksolutions.commons.web.rest.Message;
-import io.commercestacksolutions.priceproviderservice.dataaccess.organization.type.OrganizationType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.organization.definitions.OrganizationType;
 import io.commercestacksolutions.priceproviderservice.facade.organization.restentity.OrganizationRestEntity;
 import io.commercestacksolutions.priceproviderservice.web.controller.adminapi.OrganizationController;
 import io.commercestacksolutions.priceproviderservice.facade.organization.OrganizationFacade;
