+++ service/src/test/java/io/commercestacksolutions/priceproviderservice/web/controller/PriceRowControllerQueryFilterTest.java
@@ -3,7 +3,7 @@ package io.commercestacksolutions.priceproviderservice.web.controller;
 import io.commercestacksolutions.priceproviderservice.dataaccess.currency.entity.CurrencyEntity;
 import io.commercestacksolutions.priceproviderservice.dataaccess.currency.CurrencyEntityRepository;
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.entity.PriceRowEntity;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.type.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.PriceRowEntityRepository;
 import io.commercestacksolutions.priceproviderservice.dataaccess.taxclass.entity.TaxClassEntity;
 import io.commercestacksolutions.priceproviderservice.dataaccess.taxclass.TaxClassEntityRepository;
